split out galaxy specific configuration to its own config class.
